+++ ChangeLog.md
@@ -3,6 +3,7 @@ Mirrors change log
 
 ## ?.?.? / ????-??-??
 
+* Added lang.mirrors.Constants::declaredIn() - @thekid
 * Implemented `toString()` and string casts for all members: Fields,
   methods, constants and constructors.
   (@thekid)